fix: component changes
