+++ src/components/Molecules/Address/NFTApprovalList/NFTApprovalList.tsx
@@ -1,19 +1,53 @@
-import { type TokensApprovalItem } from "@covalenthq/client-sdk";
-import { type ColumnDef } from "@tanstack/react-table";
-import { TableHeaderSorting, TableList } from ".";
-import { type TokenApprovalsTableProps } from "@/utils/types/shared.types";
+import { type Option, None, Some } from "@/utils/option";
+import type { NftApprovalsItem } from "@covalenthq/client-sdk";
+import { useEffect, useState } from "react";
+import { type NFTApprovalListProps } from "@/utils/types/molecules.types";
+import { useGoldRush } from "@/utils/store";
+import { type CovalentAPIError } from "@/utils/types/shared.types";
+import { defaultErrorMessage } from "@/utils/constants/shared.constants";
+import { TableHeaderSorting, TableList } from "@/components/Shared";
+import { ColumnDef } from "@tanstack/react-table";
 import { Address } from "@/components/Atoms";
 
-export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
-    errorMessage,
-    maybeResult,
+export const NFTApprovalList: React.FC<NFTApprovalListProps> = ({
+    chain_name,
+    address,
+    ...props
 }) => {
-    const columns: ColumnDef<TokensApprovalItem>[] = [
+    const { covalentClient } = useGoldRush();
+
+    const [maybeResult, setMaybeResult] =
+        useState<Option<NftApprovalsItem[] | null>>(None);
+    const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+    useEffect(() => {
+        (async () => {
+            setMaybeResult(None);
+            setErrorMessage(null);
+            try {
+                const { data, ...error } =
+                    await covalentClient.SecurityService.getNftApprovals(
+                        chain_name,
+                        address.trim()
+                    );
+                if (error.error) {
+                    throw error;
+                }
+                setMaybeResult(new Some(data.items));
+            } catch (error: CovalentAPIError | any) {
+                setErrorMessage(error?.error_message ?? defaultErrorMessage);
+                setMaybeResult(new Some(null));
+                console.error(error);
+            }
+        })();
+    }, [chain_name, address]);
+
+    const columns: ColumnDef<NftApprovalsItem>[] = [
         {
             id: "token_details",
             accessorKey: "token_details",
             header: ({ column }) => (
-                <TableHeaderSorting<TokensApprovalItem>
+                <TableHeaderSorting<NftApprovalsItem>
                     align="left"
                     header={"Token"}
                     column={column}
@@ -23,30 +57,24 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
                 return (
                     <div className="flex flex-col">
                         <div className="flex items-center gap-1">
-                            <img
-                                src={
-                                    row.original.logo_url ||
-                                    "https://goldrush.vercel.app/icons/token.svg"
-                                }
-                                alt={row.original.ticker_symbol}
-                                className="h-6 w-6"
-                            />
-                            {row.original.ticker_symbol || (
-                                <Address address={row.original.token_address} />
+                            {row.original.contract_ticker_symbol || (
+                                <Address
+                                    address={row.original.contract_address}
+                                />
                             )}
                         </div>
                         <p className="text-xs opacity-80">
-                            <Address address={row.original.token_address} />
+                            <Address address={row.original.contract_address} />
                         </p>
                     </div>
                 );
             },
         },
         {
-            id: "balance_quote",
-            accessorKey: "balance_quote",
+            id: "token_balance",
+            accessorKey: "token_balance",
             header: ({ column }) => (
-                <TableHeaderSorting<TokensApprovalItem>
+                <TableHeaderSorting<NftApprovalsItem>
                     align="left"
                     header={"Wallet Balance"}
                     column={column}
@@ -55,35 +83,16 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
             cell: ({ row }) => {
                 return (
                     <div className="flex flex-col">
-                        <p>
-                            {Number(row.original.balance) /
-                                Math.pow(10, row.original.contract_decimals)}
-                        </p>
+                        <p>{row.original.token_balances.length}</p>
                         <p className="text-xs opacity-80">
-                            {row.original.pretty_balance_quote}
+                            ID:{" "}
+                            {row.original.token_balances.map((balance) => (
+                                <span key={balance.token_id}>
+                                    {balance.token_id?.toString()}
+                                </span>
+                            )) || "N/A"}
                         </p>
                     </div>
-                );
-            },
-        },
-        {
-            id: "pretty_value_at_risk_quote",
-            accessorKey: "pretty_value_at_risk_quote",
-            header: ({ column }) => (
-                <TableHeaderSorting<TokensApprovalItem>
-                    align="left"
-                    header={"Value at Risk"}
-                    column={column}
-                />
-            ),
-            cell: ({ row }) => {
-                return (
-                    <p>
-                        {row.original.pretty_value_at_risk_quote
-                            ? row.original.pretty_value_at_risk_quote
-                            : Number(row.original.value_at_risk) /
-                              Math.pow(10, row.original.contract_decimals)}
-                    </p>
                 );
             },
         },
@@ -91,9 +100,9 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
             id: "spender_address_label",
             accessorKey: "spender_address_label",
             header: ({ column }) => (
-                <TableHeaderSorting<TokensApprovalItem>
+                <TableHeaderSorting<NftApprovalsItem>
                     align="left"
-                    header={"Spender(s)"}
+                    header={"Spender"}
                     column={column}
                 />
             ),
@@ -118,7 +127,7 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
             id: "risk_factor",
             accessorKey: "risk_factor",
             header: ({ column }) => (
-                <TableHeaderSorting<TokensApprovalItem>
+                <TableHeaderSorting<NftApprovalsItem>
                     align="left"
                     header={"Risk Factor"}
                     column={column}
@@ -128,14 +137,12 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
                 return (
                     <span
                         className={`${
-                            row.original.spenders[0].risk_factor ===
-                            "CONSIDER REVOKING"
+                            row.original.spenders[0].allowance === "Unlimited"
                                 ? "bg-red-500"
                                 : "bg-green-500"
                         } rounded px-2 py-1 text-white`}
                     >
-                        {row.original.spenders[0].risk_factor ===
-                        "CONSIDER REVOKING"
+                        {row.original.spenders[0].allowance === "Unlimited"
                             ? "High"
                             : "Low"}
                     </span>
@@ -145,7 +152,7 @@ export const TokenApprovalsTable: React.FC<TokenApprovalsTableProps> = ({
     ];
 
     return (
-        <TableList<TokensApprovalItem>
+        <TableList<NftApprovalsItem>
             columns={columns}
             errorMessage={errorMessage}
             maybeData={maybeResult}
