+++ src/components/Molecules/index.ts
@@ -8,7 +8,8 @@ export { LatestBlocks } from "./Block/LatestBlocks/LatestBlocks";
 export { ChainSelector } from "./ChainSelector/ChainSelector";
 export { GasCard } from "./GasCard/GasCard";
 export { LatestPrice } from "./LatestPrice/LatestPrice";
-export { TokenApprovals } from "./Address/TokenApproval/TokenApprovals";
+export { TokenApprovalList } from "./Address/TokenApprovalList/TokenApprovalList";
+export { NFTApprovalList } from "./Address/NFTApprovalList/NFTApprovalList";
 export { NFTCollectionDetails } from "./NFT/NFTCollectionDetails/NFTCollectionDetails";
 export { NFTCollectionTokensList } from "./NFT/NFTCollectionTokensList/NFTCollectionTokensList";
 export { NFTFloorPrice } from "./NFT/NFTFloorPrice/NFTFloorPrice";
